Review fixes II
